important commits

+++ app/Http/Controllers/HouseController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Auth;
 use App\House;
 use App\Boarding;
+use Image;
 use Alert;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
@@ -62,7 +63,7 @@ class HouseController extends Controller
      */
     public function edit(House $house)
     {
-        if ($house->boarding->user_id == auth()->id()) {
+        if ($house->boarding->user_id == auth()->id() || Auth::guard('admin')->check()) {
 
             return view('addBoarding.editHouse',compact('house'), array('user' => Auth::user()));
         } else {
@@ -80,9 +81,122 @@ class HouseController extends Controller
      * @param  \App\House  $house
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, House $house)
+    public function update(Request $request)
     {
-        //
+        $boarding = Boarding::find(request('boardingid'));
+        $userId = $boarding->user_id;
+        $adhouse = House::find(request('houseid'));
+
+        if ( $adhouse->boarding->user_id == auth()->id() || Auth::guard('admin')->check()){
+            $request->validate([
+                'boardingType' => 'required',
+                'NoOfRooms' => 'required',
+                'Acavalability' => 'required',
+                'MonthlyRent' => 'required',
+                'KeyMoney' => 'required',
+                'Address' => 'required',
+                'Description' => '',
+                // 'Province' => 'required',
+                // 'District' => 'required',
+                // 'City' => 'required',
+                // 'Email' => 'required',
+                // 'Telephone' => 'required',
+            ]);
+
+        if($request->hasfile('filename'))
+                {
+
+                foreach($request->file('filename') as $image)
+                {
+                    $name= uniqid('real_') . '.' . $image->getClientOriginalExtension();
+                    //$image->move(public_path().'/uploads/property/house', $name);
+                    Image::make($image)->resize(1280,876)->save(\public_path('/images/uploads/boardingimg/' . $name));  
+                    $data[] = $name;
+                }
+
+                
+            }
+
+            $boarding ->user_id = $userId;
+            $boarding ->boardingType = request('boardingType');
+
+            if($request->has('School_boys')){
+                $boarding->School_boys = 1;
+            }else{
+                $boarding->School_boys = 0;
+            }
+
+            if($request->has('School_girls')){
+                $boarding->School_girls = 1;
+            }else{
+                $boarding->School_girls = 0;
+            }
+
+            if($request->has('Uni_boys')){
+                $boarding->Uni_boys = 1;
+            }else{
+                $boarding->Uni_boys = 0;
+            }
+
+            if($request->has('Uni_girls')){
+                $boarding->Uni_girls = 1;
+            }else{
+                $boarding->Uni_girls = 0;
+            }
+
+            if($request->has('Office_boys')){
+                $boarding->Office_boys = 1;
+            }else{
+                $boarding->Office_boys = 0;
+            }
+
+            if($request->has('Office_girls')){
+                $boarding->Office_girls = 1;
+            }else{
+                $boarding->Office_girls = 0;
+            } 
+
+            $boarding ->MonthlyRent = request('MonthlyRent');
+            $boarding ->KeyMoney = request('KeyMoney');
+            $boarding ->Address = request('Address');
+            $boarding ->Description = request('Description');
+            $boarding ->Province = request('Province');
+            $boarding ->District = request('District');
+            $boarding ->City = request('City');
+
+            // $boarding ->Name = request('images');
+            if ($request->hasfile('filename')) {
+
+                $boarding->filename  = json_encode($data);
+            }
+            
+            $boarding ->Email = request('Email');
+            $boarding ->Telephone = request('Telephone');
+            $boarding->save();
+
+            // $adhouse->Boarding()->associate($boarding);
+            $adhouse ->NoOfRooms = request('NoOfRooms');
+            $adhouse ->NoOfBed = request('NoOfBed');
+            $adhouse ->Acavalability = request('Acavalability');
+
+            $adhouse ->kitchenavalability = request('kitchenavalability');
+
+            if($request->has('Withfurniture')){
+                $adhouse->Withfurniture = 1;
+            }else{
+                $adhouse->Withfurniture = 0;
+            }
+
+            $adhouse->save();
+
+            if ( Auth::guard('admin')->check()) {
+                toastr()->success('User Boarding has been successfully Updated!');
+                return redirect('admin');
+            } else {
+                return redirect()->route('home');   
+            }
+            // return redirect()->route('home');   
+        }
     }
 
     /**
@@ -93,10 +207,20 @@ class HouseController extends Controller
      */
     public function destroy(House $house)
     {
-        DB::table('houses')->where('id', '=', $house->id)->delete();
-        DB::table('boardings')->where('id', '=', $house->boarding->id)->delete();
-        Alert::success('User Boarding has been deleted successfully!', 'Successfully Deleted!')->autoclose(3000);
-        return redirect()->route('home');
+        if ( $house->boarding->user_id == auth()->id() || Auth::guard('admin')->check()){
+
+            DB::table('houses')->where('id', '=', $house->id)->delete();
+            DB::table('boardings')->where('id', '=', $house->boarding->id)->delete();
+
+                       
+            if ( Auth::guard('admin')->check()) {
+                toastr()->success('User Boarding  successfully Deleted!');
+                return view('admin.index');
+            } else {
+                toastr()->success('Your Boarding  successfully Deleted!');
+                return redirect()->route('home');   
+            }
+        }
     }
     public function searchHouse(Request $request)
     {
